Add partition master and fix Glint running on Spark
Previously only one task per executor was created to distribute
parameter servers among Spark executors. When using multiple cores per
executor this did not work. So now nrOfExecutors * executorCores tasks
are created.

We further need to know which partition belongs to which partial
Word2Vec model and we also want to control the number of parameter
servers to create. A partition master distributing the available
partitions to the parameter servers solves this problem. Parameter
servers which can't obtain a partition are shut down again.