Execute dotProd and adjust operations in parallel
This results in shared mutable partial matrix state and therefore the
possibility that adjust operations overwrite each others changes.
This parallelism is however required for good performance. Training
word vectors with glint-word2vec on 53MB of data with 10 executors and
10 executor cores, I observed a runtime decrease from nearly 9 minutes
to 1 minute. According to previous research this parallelism still
results in nearly optimal convergence for sparse optimization problems.