Add parameter server cpu cores limit
The cpu cores per parameter server can be limited to not use more cores
than requested by the Spark application. The cores now are determined
by the Spark application in which the parameter servers run and not the
one creating the partial matrices. All partial matrices on a parameter
server also share these cores now.